backend-api/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
asyncpg==0.30.0
pydantic==2.11.3
pydantic-settings==2.8.1
python-jose[cryptography]==3.4.0
python-multipart==0.0.20
python-dotenv==1.1.0
groq==0.23.1
whisper-ctranslate2==0.1.1
